cde平台告警清理

+++ interfaces/kits/js/src/mod_fs/properties/copy_listener/trans_listener.cpp
@@ -176,7 +176,8 @@ void TransListener::CallbackComplete(uv_work_t *work, int stat)
         return;
     }
     NVal obj = NVal::CreateObject(env);
-    if (entry->progressSize <= numeric_limits<int64_t>::max() && entry->totalSize <= numeric_limits<int64_t>::max()) {
+    if (entry->progressSize <= numeric_limits<int64_t>::max() &&
+        entry->totalSize <= static_cast<uint64_t>(numeric_limits<int64_t>::max())) {
         obj.AddProp("processedSize", NVal::CreateInt64(env, entry->progressSize).val_);
         obj.AddProp("totalSize", NVal::CreateInt64(env, entry->totalSize).val_);
     }
